requirements.txt
pydantic-settings==2.12.0
pydantic==2.12.5
uvicorn==0.40.0
fastapi==0.128.0
boto3==1.42.21
python-jose==3.5.0
passlib==1.7.4